+++ log-viewer-app/src/main/resources/io/github/qupath/logviewer/app/strings.properties
@@ -1,5 +1,3 @@
-title = Log viewer
-
 # File menu
 Menu.File = File
 Action.File.close = Close
