+++ log-viewer-app/src/main/resources/io/github/qupath/logviewer/app/strings_fr.properties
@@ -1,5 +1,3 @@
-title = Visionneuse de journaux
-
 # File menu
 Menu.File = Fichier
 Action.File.close = Fermer
